+++ evtloop.h
@@ -43,6 +43,8 @@ typedef struct {
 	uint64_t el_radar39267328;
 	uint64_t el_radar39623812_fatal;
 	uint64_t el_radar39623812;
+	uint64_t el_radar42770257_fatal;
+	uint64_t el_radar42770257;
 	uint64_t el_missingtoken;
 	uint64_t el_ooms;
 	aupipe_stat_t ap;